Add support for marking orders Shipped (Tracked/Untracked)
Finally implement the "Mark Shipped" option.
For untracked orders it will just make the API request and if it
succeeds, it marks the order shipped.
For tracked orders it will show a dialog where the tracking number
must be provided. A Settings option is also added so a URL template
can be provided and if it is provided, it will be auto-filled in the
dialog.